Update target_working_minutes.json
(I commited this to my old branch yesterday but I think it is better to have it all in one branch, i will delete the old branch)

Updated target_working_minutes.json with the right work time for each employee. The "target" represents the target working minutes per employee for the month.

Additionally added a "actual" attribute which represents the already "worked" time of the employee (absences increase the "IST" time instead of reducing the "SOLL" time). Therefore it is necessary to subtract the actual from the target time to get the time that is left for each employee for the month.